Separates yaml conversion for ValueRanges in separate YamlProtocol; includes additional unit tests
